Move Scala-2 DI specific tests to legacy package
