+++ airframe/src/test/scala-2/wvlet/airframe/legacy/ChildSessionTest.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe
+package wvlet.airframe.legacy
 import java.util.concurrent.atomic.{AtomicBoolean, AtomicInteger}
 
 import wvlet.airspec.AirSpec
