+++ airframe/src/test/scala-2/wvlet/airframe/legacy/lifecycle/LazyStartTest.scala
